# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Event Hub API",
        "description": "API para gerenciamento de eventos com venda de ingressos, processamento de pagamentos e notificações.",
        "contact": {
            "email": "contato@eventhub.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/v1",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/buy-ticket": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Comprar ingressos",
                "description": "Realiza a compra de ingressos para um evento",
                "operationId": "da991c0e1abe7d76291de3fef5a40fdc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "event_id",
                                    "quantity",
                                    "card_number",
                                    "card_holder_name",
                                    "card_expiration_date",
                                    "card_cvv"
                                ],
                                "properties": {
                                    "event_id": {
                                        "type": "string",
                                        "example": "ulid"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "discount_coupon": {
                                        "type": "string",
                                        "example": "BLACKFRIDAY"
                                    },
                                    "card_number": {
                                        "type": "string",
                                        "example": "1234567890123456"
                                    },
                                    "card_holder_name": {
                                        "type": "string",
                                        "example": "JOAO SILVA"
                                    },
                                    "card_expiration_date": {
                                        "type": "string",
                                        "example": "12/25"
                                    },
                                    "card_cvv": {
                                        "type": "string",
                                        "example": "123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pedido criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "ulid"
                                        },
                                        "event_id": {
                                            "type": "string",
                                            "example": "ulid"
                                        },
                                        "participant_id": {
                                            "type": "string",
                                            "example": "ulid"
                                        },
                                        "quantity": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "ticket_price": {
                                            "type": "number",
                                            "example": 50
                                        },
                                        "discount": {
                                            "type": "number",
                                            "example": 5
                                        },
                                        "total_amount": {
                                            "type": "number",
                                            "example": 95
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "CONFIRMED"
                                        },
                                        "tickets": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "ulid"
                                                    },
                                                    "order_id": {
                                                        "type": "string",
                                                        "example": "ulid"
                                                    },
                                                    "event_id": {
                                                        "type": "string",
                                                        "example": "ulid"
                                                    },
                                                    "participant_id": {
                                                        "type": "string",
                                                        "example": "ulid"
                                                    },
                                                    "used_at": {
                                                        "type": "string",
                                                        "example": null,
                                                        "nullable": true
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2025-10-20 12:00:00"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2025-10-20 12:00:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2025-10-20 12:00:00"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2025-10-20 12:00:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "422": {
                        "description": "Dados de validação inválidos"
                    },
                    "400": {
                        "description": "Erro no processamento do pedido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Listar pedidos do usuário",
                "description": "Retorna todos os pedidos do usuário autenticado",
                "operationId": "5f78692ecf363eff1a55a9fae908dbf5",
                "responses": {
                    "200": {
                        "description": "Lista de pedidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "ulid"
                                                    },
                                                    "event_id": {
                                                        "type": "string",
                                                        "example": "ulid"
                                                    },
                                                    "participant_id": {
                                                        "type": "string",
                                                        "example": "ulid"
                                                    },
                                                    "quantity": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "ticket_price": {
                                                        "type": "number",
                                                        "example": 50
                                                    },
                                                    "discount": {
                                                        "type": "number",
                                                        "example": 5
                                                    },
                                                    "total_amount": {
                                                        "type": "number",
                                                        "example": 95
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "CONFIRMED"
                                                    },
                                                    "tickets": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "string",
                                                                    "example": "ulid"
                                                                },
                                                                "order_id": {
                                                                    "type": "string",
                                                                    "example": "ulid"
                                                                },
                                                                "event_id": {
                                                                    "type": "string",
                                                                    "example": "ulid"
                                                                },
                                                                "participant_id": {
                                                                    "type": "string",
                                                                    "example": "ulid"
                                                                },
                                                                "used_at": {
                                                                    "type": "string",
                                                                    "example": null,
                                                                    "nullable": true
                                                                },
                                                                "created_at": {
                                                                    "type": "string",
                                                                    "example": "2025-10-20 12:00:00"
                                                                },
                                                                "updated_at": {
                                                                    "type": "string",
                                                                    "example": "2025-10-20 12:00:00"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2025-10-20 12:00:00"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2025-10-20 12:00:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Buscar pedido por ID",
                "description": "Retorna os detalhes de um pedido específico",
                "operationId": "637dfd376085208002bf92ac5462e42c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do pedido",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do pedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "ulid"
                                        },
                                        "event_id": {
                                            "type": "string",
                                            "example": "ulid"
                                        },
                                        "participant_id": {
                                            "type": "string",
                                            "example": "ulid"
                                        },
                                        "quantity": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "ticket_price": {
                                            "type": "number",
                                            "example": 50
                                        },
                                        "discount": {
                                            "type": "number",
                                            "example": 5
                                        },
                                        "total_amount": {
                                            "type": "number",
                                            "example": 95
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "CONFIRMED"
                                        },
                                        "tickets": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "ulid"
                                                    },
                                                    "order_id": {
                                                        "type": "string",
                                                        "example": "ulid"
                                                    },
                                                    "event_id": {
                                                        "type": "string",
                                                        "example": "ulid"
                                                    },
                                                    "participant_id": {
                                                        "type": "string",
                                                        "example": "ulid"
                                                    },
                                                    "used_at": {
                                                        "type": "string",
                                                        "example": null,
                                                        "nullable": true
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2025-10-20 12:00:00"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2025-10-20 12:00:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2025-10-20 12:00:00"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2025-10-20 12:00:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Pedido não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticação via token Bearer",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}